Add Signal type for values stored outside of a Binding
Sometimes you want to store a value somewhere else (not in a `Binding`).
In that case, you can create a `Signal` that the value has been changed.
Then each `ComputedBinding` which uses the value
must `observe()` the `Signal`.
Additionally, when the value is changed, the mutating function must call
`Signal::emit()` as well.
If either of these is forgotten, you *will* get desyncs.